updated dafi_pipeline for GP to include skipping preprocessing

+++ GP/resources/dafi_pipeline.sh
@@ -4,6 +4,7 @@ usage="$(basename "$0") [--help] [-s n] -- DAFi pipeline script to transform, ga
 
 where:
     --help  show this help text
+    --skip_prep   skips preprocessing and header rearrangement
     --no_opt  disables intel AVX2 level optimization in case of compatibility issues (default: AVX2 enabled if detected)
     --nclusters specify number of initial clusters to use for the k-means clustering on the whole sample (default: 100)
     --nreclusters specify number of clusters to use for the re-clustering after applying filtering on the parent population (default: same as nclusters)
@@ -12,6 +13,7 @@ where:
 
 if grep -q avx2 /proc/cpuinfo; then optimize=true; else optimize=false; fi
 
+skipPrep=false
 numOfClusters=100
 numOfReClusters=$numOfClusters
 numOfCores=$(grep -c ^processor /proc/cpuinfo)
@@ -19,6 +21,7 @@ seed=2
 
 while [[ "$#" > 1 ]]; do case $1 in
     --help) echo $usage; exit;;
+    --skip_prep) skipPrep=true;;
     --no_opt) optimize=false;;
     --nclusters) numOfClusters="$2";;
     --nreclusters) numOfReClusters="$2";;
@@ -32,36 +35,46 @@ work=/home/jovyan/work
 cwd=$(pwd)
 Label=${PWD##*/}
 
-#Transformation and compensation using FCSTrans
-#Convert FCS binary to txt format 
-if [ -d "TXT" ]; then
-        rm -Rf TXT
-        mkdir TXT
-else
-        mkdir TXT
-fi
-cd TXT
-run_FCSTrans2TXT.R $work/FCS
-cd ..
 
-#Arrange columns and replace labels
-if [ -d "Preprocessed" ]; then
-        rm -Rf Preprocessed
-        mkdir Preprocessed
+if [ "$skipPrep" != true ]; then
+	#Transformation and compensation using FCSTrans
+	#Convert FCS binary to txt format 
+	if [ -d "TXT" ]; then
+        	rm -Rf TXT
+        	mkdir TXT
+	else
+        	mkdir TXT
+	fi
+	cd TXT
+	run_FCSTrans2TXT.R $work/FCS
+	cd ..
+
+	#Arrange columns and replace labels
+	if [ -d "Preprocessed" ]; then
+        	rm -Rf Preprocessed
+        	mkdir Preprocessed
+	else
+        	mkdir Preprocessed
+	fi
+	cd Preprocessed
+	count=1
+	for f in ../TXT/*.txt
+	do
+	filename=$(basename "$f");
+	(ArrangeHeader.awk $work/config/header.lst $work/config/header.txt $f > $filename)&
+	if [ $(($count % $numOfCores)) -eq 0 ]; then wait; fi
+	count=$((count+1));
+	done
+	wait
+	cd ..
 else
-        mkdir Preprocessed
+	if [ -d "Preprocessed" ]; then
+		echo "Found Preprocessed Data, Skipping preprocessing"
+	else
+		echo "No Preprocessed dir found! Aborting!"
+		exit 1
 fi
-cd Preprocessed
-count=1
-for f in ../TXT/*.txt
-do
-filename=$(basename "$f");
-(ArrangeHeader.awk $work/config/header.lst $work/config/header.txt $f > $filename)&
-if [ $(($count % $numOfCores)) -eq 0 ]; then wait; fi
-count=$((count+1));
-done
-wait
-cd ..
+
 cp $work/config/inclusion.config pipeline.config
 
 #Gating and filtering via DAFi framework
